E5 fix dtos input/output fuente dinamica

+++ src/main/java/ar/utn/frba/ddsi/cliente_liviano/models/Categoria.java
@@ -6,11 +6,11 @@ import lombok.Setter;
 @Setter
 @Getter
 public class Categoria {
-    public Integer id;
+    public Long id;
     private String nombre;
 
     public Categoria() {}
-    public Categoria(Integer id, String nombre) {
+    public Categoria(Long id, String nombre) {
         this.id = id;
         this.nombre = nombre;
     }
